+++ docs/.vitepress/config.ts
@@ -10,44 +10,46 @@ export default defineConfig({
 		},
 		nav: [
 			{ text: "Home", link: "/" },
-			{ text: "Guide", link: "/guide/installation" },
+			{ text: "Guide", link: "/guide/getting-started" },
 			{ text: "Maintained by Digitalastronaut", link: "https://digitalastronaut.be/" },
 		],
 
 		sidebar: [
 			{
 				text: "Introduction",
+				link: "/guide/motivation",
 				items: [
-					{ text: "Why this toolbar?", link: "/guide/why" },
+					{ text: "About", link: "/guide/motivation" },
 					{ text: "Installation guide", link: "/guide/installation" },
 					{ text: "Gettings started", link: "/guide/getting-started" },
 				],
 			},
 			{
-				text: "Features",
+				text: "Toolbar chunks",
+				link: "/chunks/",
 				items: [
-					{ text: "Start menu", link: "/features/start-menu" },
-					{ text: "Quick actions", link: "/features/quick-actions" },
-					{ text: "Global search", link: "/features/global-search" },
+					{ text: "Start menu", link: "/chunks/start-menu" },
+					{ text: "Quick actions", link: "/chunks/quick-actions" },
+					{ text: "Global search", link: "/chunks/global-page-search" },
 					{
 						text: "SEO menu",
-						link: "/features/seo-menu",
+						link: "/chunks/seo-menu",
 						items: [
-							{ text: "Base checklist", link: "/features/seo-base-checklist" },
-							{ text: "Share previews", link: "/features/seo-base-checklist" },
-							{ text: "Image analytics", link: "/features/seo-base-checklist" },
-							{ text: "Headings hierarchy", link: "/features/seo-base-checklist" },
-							{ text: "Structured data", link: "/features/seo-base-checklist" },
-							{ text: "Base checklist", link: "/features/seo-base-checklist" },
+							{ text: "Base checklist", link: "/chunks/seo-menu/base-checklist" },
+							{ text: "Share previews", link: "/chunks/seo-menu/share-previews" },
+							{ text: "Image analytics", link: "/chunks/seo-menu/image-analytics" },
+							{ text: "Headings", link: "/chunks/seo-menu/headings" },
+							// { text: "Structured data", link: "/chunks/seo-menu/structured-data" },
 						],
 					},
+					{ text: "Help menu", link: "/chunks/help-menu" },
 				],
 			},
 			{
 				text: "Extending the toolbar",
 				items: [
-					{ text: "Extra start menu links", link: "/extending/extra-start-menu-links" },
-					{ text: "Custom widgets", link: "/extending/custom-widgets" },
+					// { text: "Extra start menu links", link: "/extending/extra-start-menu-links" },
+					{ text: "Custom chunks", link: "/extending/custom-chunks" },
 				],
 			},
 		],